% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorfun.R
\name{addShade}
\alias{addShade}
\title{Add shade to a color name}
\usage{
addShade(name, shade = 0)
}
\arguments{
\item{name}{name of the color}

\item{shade}{shade value (between 0 and 1). For zero, the same color is returned.}
}
\value{
a new color name
}
\description{
Add shade to a color name
}
\author{
Christoffer Moesgaard Albertsen
}
