% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorfun.R
\name{addTint}
\alias{addTint}
\title{Add tint to a color name}
\usage{
addTint(name, tint = 0)
}
\arguments{
\item{name}{name of the color}

\item{tint}{tint value (between 0 and 1). For 1, the same}
}
\value{
a new color name
}
\description{
Add tint to a color name
}
\author{
Christoffer Moesgaard Albertsen
}
