% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorfun.R
\name{addTone}
\alias{addTone}
\title{Add tone to a color name}
\usage{
addTone(name, tone = 0)
}
\arguments{
\item{name}{name of the color}

\item{tone}{tone value (between 0 and 1). For 0, the same color is returned.}
}
\value{
a new color name
}
\description{
Add tone to a color name
}
\author{
Christoffer Moesgaard Albertsen
}
