% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorfun.R
\name{addTrans}
\alias{addTrans}
\title{Add alpha value to a color name}
\usage{
addTrans(name, alpha = 1)
}
\arguments{
\item{name}{name of the color}

\item{alpha}{alpha value (between 0 and 1)}
}
\value{
a new color name
}
\description{
Add alpha value to a color name
}
\author{
Christoffer Moesgaard Albertsen
}
