% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation_convenience.R
\name{axisInside}
\alias{axisInside}
\title{Draw figure axis inside plot}
\usage{
axisInside(side)
}
\arguments{
\item{side}{side of plot to draw on}
}
\value{
Nothing, but plots as side effect
}
\description{
Draw figure axis inside plot
}
\author{
Christoffer Moesgaard Albertsen
}
