% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beanplot.R
\name{beanplot}
\alias{beanplot}
\title{Beanplot of something}
\usage{
beanplot(x, ...)
}
\arguments{
\item{x}{Some object}

\item{...}{Other arguments}
}
\value{
Creates a plot
}
\description{
Beanplot of something
}
\author{
Christoffer Moesgaard Albertsen
}
