% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation_convenience.R
\name{bgtext}
\alias{bgtext}
\title{Add text with background to plot}
\usage{
bgtext(
  x,
  y,
  labels,
  ...,
  cex = 1,
  font = NULL,
  bg = "white",
  bgex = 1,
  border = NA
)
}
\arguments{
\item{x}{x-coordinate}

\item{y}{y-coordinate}

\item{labels}{label to add}

\item{...}{arguments passed to text}

\item{cex}{size of text}

\item{font}{font of text}

\item{bg}{background color}

\item{bgex}{background extend factor}

\item{border}{border color of background}
}
\description{
Add text with background to plot
}
\author{
Christoffer Moesgaard Albertsen
}
