% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_from_github.R
\name{buildFromGithub}
\alias{buildFromGithub}
\title{Download and build an R package from GitHub}
\usage{
buildFromGithub(repo, ref = "master", subdir = NULL,
  buildArgs = c("--no-build-vignettes"))
}
\arguments{
\item{repo}{GitHub user and repository separated by /}

\item{ref}{Reference to commit or branch. Default is master}

\item{subdir}{Path to subdir containing the package. Should be NULL if the package is in the top directory}

\item{buildArgs}{Character vector of arguments passed to R CMD build. Default is "--no-build-vignettes"}
}
\value{
Nothing
}
\description{
Download and build an R package from GitHub
}
\author{
Christoffer Moesgaard Albertsen
}
