% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation_convenience.R
\name{collapse}
\alias{collapse}
\title{Collapse vector to string}
\usage{
collapse(
  x,
  collap = c(rep(", ", length(x) - 2), ifelse(length(x) > 2, ", and ", " and "))
)
}
\arguments{
\item{x}{vector of values}

\item{collap}{rule for collapsing}
}
\value{
a string
}
\description{
Collapse vector to string
}
\author{
Christoffer Moesgaard Albertsen
}
