% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation_convenience.R
\name{displayNum}
\alias{displayNum}
\title{Convert number to display text}
\usage{
displayNum(
  x,
  capitalize = FALSE,
  big.mark = ",",
  decimal.mark = ".",
  small.mark = "",
  digits = 0
)
}
\arguments{
\item{x}{number}

\item{capitalize}{Capitalize first letter?}
}
\value{
LaTeX code for the fraction
}
\description{
Convert number to display text
}
\author{
Text representation of number
}
