% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation_convenience.R
\name{dtucols}
\alias{dtucols}
\title{DTU color palette}
\usage{
dtucols(x)
}
\arguments{
\item{x}{name of colors to return}
}
\value{
color codes
}
\description{
DTU color palette
}
\author{
Christoffer Moesgaard Albertsen
}
