% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dz2pdf.R
\name{dz2pdf}
\alias{dz2pdf}
\title{Convert DZSlides to PDF}
\usage{
dz2pdf(filein, fileout, sizeFactor = 4, delay = 0.5)
}
\arguments{
\item{filein}{Path and name of input file with DZSlides. Should either be in working directory or be on the form "file:///home/full/path/to/slides.html"}

\item{fileout}{Name of output pdf file}

\item{sizeFactor}{Factor to scale png of slides}

\item{delay}{Delay before capturing slide}
}
\value{
Nothing
}
\description{
Uses the webshot package to convert html DZSlides to a PDF (through PNG images).
}
\author{
Christoffer Moesgaard Albertsen
}
