% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dzslides_presentation.R
\name{dzslides_presentation}
\alias{dzslides_presentation}
\title{Convert to a DZSlides presentation}
\usage{
dzslides_presentation(incremental = FALSE, fig_width = 8, fig_height = 6,
  fig_retina = if (!fig_caption) 2, fig_caption = FALSE, dev = "png",
  smart = TRUE, self_contained = TRUE, highlight = "default",
  mathjax = "default", template = "default", css = NULL,
  includes = NULL, keep_md = FALSE, lib_dir = NULL,
  md_extensions = NULL, pandoc_args = NULL, ...)
}
\value{
R Markdown output format to pass to \code{\link{render}}
}
\description{
Format for converting from R Markdown to a DZSlides presentation.
}
\details{
See the \href{http://rmarkdown.rstudio.com/slidy_presentation_format.html}{online
documentation} for additional details on using the \code{slidy_presentation} format.

For more information on markdown syntax for presentations see
\href{http://johnmacfarlane.net/pandoc/demo/example9/producing-slide-shows-with-pandoc.html}{producing
slide shows with pandoc}.
}
\examples{
\dontrun{

library(rmarkdown)

# simple invocation
render("pres.Rmd", dzslides_presentation())

# specify an option for incremental rendering
render("pres.Rmd", dzslides_presentation(incremental = TRUE))
}
}

