% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation_convenience.R
\name{fd}
\alias{fd}
\title{Format number}
\usage{
fd(x, digits = 1)
}
\arguments{
\item{x}{number}

\item{digits}{number of digits}
}
\value{
string
}
\description{
Small wrapper for formatC to limit text needed
}
\author{
Christoffer Moesgaard Albertsen
}
