% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gcov.R
\name{gcovTable}
\alias{gcovTable}
\title{Create coverage table from list of gcov objects}
\usage{
gcovTable(x, ...)
}
\arguments{
\item{x}{list of gcov objects}

\item{...}{additional arguments passed to formatC}
}
\value{
A coverage table
}
\description{
Create coverage table from list of gcov objects
}
\author{
Christoffer Moesgaard Albertsen
}
