% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorblind.R
\name{greenBlindness}
\alias{greenBlindness}
\title{Emulate green blindness in colors}
\usage{
greenBlindness(red, green, blue, alpha = 1, names = NULL, maxColorValue = 255)
}
\arguments{
\item{red}{Red RGB value (or color HEX code or name - alpha is ignored)}

\item{green}{Green RGB value}

\item{blue}{Blue RGB value}

\item{alpha}{alpha value to use}

\item{names}{Names for the resulting vector}

\item{maxColorValue}{Maximum color value for red, green, blue}
}
\value{
New colors
}
\description{
Based on https://personal.sron.nl/~pault/data/colourschemes.pdf
}
\author{
Christoffer Moesgaard Albertsen
}
