% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagePlot.R
\name{imagePlot}
\alias{imagePlot}
\title{Plot image from file}
\usage{
imagePlot(
  x,
  objectFit = c("fill", "contain", "cover", "none", "scale-down"),
  halign = c("c", "l", "r"),
  valign = c("c", "t", "b"),
  maxWidthPct = 1,
  maxHeightPct = 1,
  add = FALSE,
  noMargin = TRUE,
  ...
)
}
\arguments{
\item{x}{path to image}

\item{objectFit}{How the image should fill the plot: "fill", "contain", "cover", "none","scale-down"}

\item{halign}{Horizontal alignment: c, l, r}

\item{valign}{Vertical alignment: c, t, b}

\item{noMargin}{Plot with oma and mar set to zero}

\item{...}{other arguments}
}
\value{
Plots the image
}
\description{
Plot image from file
}
\author{
Christoffer Moesgaard Albertsen
}
