% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_from_github.R
\name{installDependencies}
\alias{installDependencies}
\title{Download, build, and install package dependencies}
\usage{
installDependencies(descriptionPath, buildArgs = c("--no-build-vignettes"),
  installArgs = c(), dependencies = c("Depends", "Imports", "LinkingTo"))
}
\arguments{
\item{descriptionPath}{Path to a DESCRIPTION file}

\item{buildArgs}{Character vector of arguments passed to R CMD build.}

\item{installArgs}{Character vector of arguments passed to R CMD INSTALL.}

\item{dependencies}{Character vector of dependency types to install ("Depends","Imports","LinkingTo","Enhances", or "Suggests")}
}
\value{
Nothing
}
\description{
Download, build, and install package dependencies
}
\author{
Christoffer Moesgaard Albertsen
}
