% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_from_github.R
\name{installFromGithub}
\alias{installFromGithub}
\title{Download, build, and install an R package from GitHub}
\usage{
installFromGithub(repo, ref = "master", subdir = NULL,
  buildArgs = c("--no-build-vignettes"), installArgs = c(),
  dependencies = c("Depends", "Imports", "LinkingTo"))
}
\arguments{
\item{repo}{GitHub user and repository separated by /}

\item{ref}{Reference to commit or branch. Default is master}

\item{subdir}{Path to subdir containing the package. Should be NULL if the package is in the top directory}

\item{buildArgs}{Character vector of arguments passed to R CMD build.}

\item{installArgs}{Character vector of arguments passed to R CMD INSTALL.}

\item{dependencies}{Character vector of dependency types to install ("Depends","Imports","LinkingTo","Enhances", or "Suggests")}
}
\value{
Nothing
}
\description{
Download, build, and install an R package from GitHub
}
\author{
Christoffer Moesgaard Albertsen
}
