% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation_convenience.R
\name{makeMapLegend}
\alias{makeMapLegend}
\title{Add map legend for choropleth map}
\usage{
makeMapLegend(x, cols, txt)
}
\arguments{
\item{x}{Variable values plotted}

\item{cols}{Colors}

\item{txt}{Title / description}
}
\description{
Add map legend for choropleth map
}
\author{
Christoffer Moesgaard Albertsen
}
