% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation_convenience.R
\name{makeSquare}
\alias{makeSquare}
\title{Make a pixel matrix/array square}
\usage{
makeSquare(p, value = 0)
}
\arguments{
\item{p}{Pixel matrix/array}

\item{value}{Value of added pixels}
}
\description{
Make a pixel matrix/array square
}
\author{
Christoffer Moesgaard Albertsen
}
