% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_skeleton.R
\name{packageSkeleton}
\alias{packageSkeleton}
\title{Create a package skeleton}
\usage{
packageSkeleton(pkg, path = file.path("."))
}
\arguments{
\item{pkg}{Name of the new package}

\item{path}{(existing) Directory of package}
}
\value{
Used for side effects
}
\description{
Create a package skeleton
}
\author{
Christoffer Moesgaard Albertsen
}
