% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gcov.R
\name{read_gcov}
\alias{read_gcov}
\title{Read content of .gcov file}
\usage{
read_gcov(file)
}
\arguments{
\item{file}{path to .gcov file}
}
\value{
a gcov object
}
\description{
Read content of .gcov file
}
\author{
Christoffer Moesgaard Albertsen
}
