% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_massif.R
\name{read_massif}
\alias{read_massif}
\title{Read valgrind Massif output}
\usage{
read_massif(file, keep_details = FALSE)
}
\arguments{
\item{file}{file to read}

\item{keep_details}{keep details?}
}
\value{
massif S3 object
}
\description{
Read valgrind Massif output
}
\author{
Christoffer Moesgaard Albertsen
}
