% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarfun.R
\name{solarPosition}
\alias{solarPosition}
\title{Calculate solar angles}
\usage{
solarPosition(date, lat, lon)
}
\arguments{
\item{date}{date (UTC) in the format "YYY-mm-dd HH:MM:SS"}

\item{lat}{latitude of observer}

\item{lon}{longitude of observer}
}
\value{
list of values
}
\description{
Calculate solar angles
}
\author{
Christoffer Moesgaard Albertsen
Modified from https://doi.org/10.1016/j.renene.2021.03.047
}
