% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation_convenience.R
\name{tofrac}
\alias{tofrac}
\title{Print number as (LaTeX) fraction}
\usage{
tofrac(x, dollar = TRUE)
}
\arguments{
\item{x}{number}

\item{dollar}{}
}
\value{
LaTeX code for the fraction
}
\description{
Print number as (LaTeX) fraction
}
\author{
Christoffer Moesgaard Albertsen
}
