% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorfun.R
\name{toGreyscale}
\alias{toGreyscale}
\title{Turn pixel into grey scale}
\usage{
toGreyscale(p)
}
\arguments{
\item{p}{Pixel matrix/array}
}
\value{
A grey scale pixel matrix/array
}
\description{
Turn pixel into grey scale
}
\author{
Christoffer Moesgaard Albertsen
}
