% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dzslides_presentation.R
\name{dzslides_presentation}
\alias{dzslides_presentation}
\title{Convert to a DZSlides presentation}
\usage{
dzslides_presentation(incremental = FALSE, fig_width = 8,
  fig_height = 6, fig_retina = if (!fig_caption) 2,
  fig_caption = FALSE, dev = "png", smart = TRUE,
  self_contained = TRUE, highlight = "default", mathjax = "default",
  template = "default", css = NULL, includes = NULL,
  keep_md = FALSE, lib_dir = NULL, md_extensions = NULL,
  pandoc_args = NULL, ...)
}
\arguments{
\item{incremental}{should lists be incremental?}

\item{fig_width}{figure width}

\item{fig_height}{figure height}

\item{fig_retina}{retina figure?}

\item{fig_caption}{figure caption?}

\item{dev}{figure device}

\item{smart}{smart?}

\item{self_contained}{self contained?}

\item{highlight}{highlighting style}

\item{mathjax}{mathjax}

\item{template}{template path}

\item{css}{extra css}

\item{includes}{extra includes}

\item{keep_md}{keep markdown file?}

\item{lib_dir}{...}

\item{md_extensions}{markdown extensions to use}

\item{pandoc_args}{extra pandoc arguments}

\item{...}{other arguments}
}
\value{
R Markdown output format to pass to \code{\link{render}}
}
\description{
Format for converting from R Markdown to a DZSlides presentation.
}
\examples{
\dontrun{

library(rmarkdown)

# simple invocation
render("pres.Rmd", dzslides_presentation())

# specify an option for incremental rendering
render("pres.Rmd", dzslides_presentation(incremental = TRUE))
}
##' @export
}
\author{
Christoffer Moesgaard Albertsen
}
